WA-232 Refactor code avoiding DRY
